+++ apps/melding-form/src/app/(general)/Home.tsx
@@ -9,7 +9,7 @@ import { InvalidFormAlert } from '@meldingen/ui'
 import type { FormState } from '../../types'
 
 import { SystemErrorAlert } from './_components/SystemErrorAlert'
-import { getDocumentTitleOnError } from './_utils/getDocumentTitleOnError'
+import { getDocumentTitleOnError } from './_utils/validation/getDocumentTitleOnError'
 
 const initialState: FormState = {}
