+++ apps/melding-form/src/app/(general)/_utils/validation/hasValidationErrors.ts
