+++ apps/melding-form/src/app/(general)/aanvullende-vragen/[classificationId]/[panelId]/AdditionalQuestions.tsx
@@ -11,8 +11,8 @@ import { InvalidFormAlert } from '@meldingen/ui'
 import type { FormState, ValidationError } from 'apps/melding-form/src/types'
 
 import { SystemErrorAlert } from '../../../_components/SystemErrorAlert'
-import { AnswersByKey } from '../../../_utils/getAnswersByKey'
-import { getDocumentTitleOnError } from '../../../_utils/getDocumentTitleOnError'
+import { AnswersByKey } from '../../../_utils/conditions/getFilteredAnswersByKey'
+import { getDocumentTitleOnError } from '../../../_utils/validation/getDocumentTitleOnError'
 import { BackLink } from 'apps/melding-form/src/app/_components'
 
 const getPrefilledFormComponents = (components: Component[], formData: FormData): Component[] =>
