+++ apps/melding-form/src/app/(general)/aanvullende-vragen/[classificationId]/[panelId]/page.tsx
@@ -14,7 +14,7 @@ import type {
 
 import { getFormClassificationByClassificationId, getMeldingByMeldingIdAnswersMelder } from '@meldingen/api-client'
 
-import { getFilteredAnswersByKey, isPanelComponentOutput } from '../../../_utils/getAnswersByKey'
+import { getFilteredAnswersByKey, isPanelComponentOutput } from '../../../_utils/conditions/getFilteredAnswersByKey'
 import { AFTER_ADDITIONAL_QUESTIONS_PATH, getPreviousPanelPath } from './_utils/navigationUtils'
 import { setDateComponentOptions } from './_utils/setDateComponentOptions'
 import { postForm } from './actions'
