+++ apps/melding-form/src/app/(general)/actions.ts
@@ -8,7 +8,7 @@ import { patchMeldingByMeldingIdMelder, postMelding } from '@meldingen/api-clien
 import { COOKIES } from '../../constants'
 import { handleApiError } from '../../handleApiError'
 import { resolveClassificationRedirect } from '../utils'
-import { hasValidationErrors } from './_utils/hasValidationErrors'
+import { hasValidationErrors } from './_utils/validation/hasValidationErrors'
 
 export type ArgsType = {
   existingId?: string
