+++ apps/melding-form/src/app/(general)/bijlage/Attachments.tsx
@@ -19,7 +19,7 @@ import type { FileUpload as FileUploadType, PendingFileUpload } from './utils'
 import type { FormState } from 'apps/melding-form/src/types'
 
 import { SystemErrorAlert } from '../_components/SystemErrorAlert'
-import { getDocumentTitleOnError } from '../_utils/getDocumentTitleOnError'
+import { getDocumentTitleOnError } from '../_utils/validation/getDocumentTitleOnError'
 import { BackLink } from '../../_components'
 import { submitAttachmentsForm } from './actions'
 import { startUpload } from './utils'
