+++ apps/melding-form/src/app/(general)/contact/actions.ts
@@ -6,7 +6,7 @@ import { redirect } from 'next/navigation'
 
 import { patchMeldingByMeldingIdContact, putMeldingByMeldingIdAddContactInfo } from '@meldingen/api-client'
 
-import { hasValidationErrors } from '../_utils/hasValidationErrors'
+import { hasValidationErrors } from '../_utils/validation/hasValidationErrors'
 import { COOKIES, TOP_ANCHOR_ID } from 'apps/melding-form/src/constants'
 import { isApiErrorArray } from 'apps/melding-form/src/handleApiError'
